GNG-418: Move Search tabMap configuration into TabMenu parent directory
